fix: tiling of IDEA context menu
The IDEA context menu should not be tiled as it reshapes the window.
The fix is to make any IDEA window without a title floating. In
addition, the rules are processed in order so I removed the default
tile on all IDEA windows to prevent going from tiling to floating.